refactor: Restructure project to internal/cmd/<resource>/ pattern
This is a major restructuring to align with CLI guide best practices:

- Move main.go to cmd/slack-cli/main.go
- Create internal/cmd/root/root.go with root command
- Split each command file into internal/cmd/<resource>/ directory
  - channels: list, get, create, archive, unarchive, set_topic, set_purpose, invite
  - users: list, get
  - messages: send, update, delete, history, thread, react, unreact
  - workspace: info
  - config: set_token, delete_token, show
- Create internal/output package for shared output flag to avoid import cycles
- Update Makefile to build from new location
- Update .goreleaser.yaml with new main path and ldflags
- Update .gitignore to allow cmd/slack-cli/ directory

All tests pass and lint is clean.
