+++ .gitignore
@@ -1,5 +1,6 @@
 # Binary
-slack-cli
+/slack-cli
+bin/
 
 # Build output
 dist/
