+++ Makefile
@@ -14,7 +14,7 @@ DIST_DIR = dist
 all: build
 
 build:
-	go build $(LDFLAGS) -o bin/$(BINARY) .
+	go build $(LDFLAGS) -o bin/$(BINARY) ./cmd/slack-cli
 
 test:
 	go test -v -race ./...
@@ -48,22 +48,22 @@ release: clean
 	mkdir -p $(DIST_DIR)
 
 	# macOS ARM64
-	GOOS=darwin GOARCH=arm64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) .
+	GOOS=darwin GOARCH=arm64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) ./cmd/slack-cli
 	tar -czvf $(DIST_DIR)/$(BINARY)_$(VERSION)_darwin_arm64.tar.gz -C $(DIST_DIR) $(BINARY)
 	rm $(DIST_DIR)/$(BINARY)
 
 	# macOS AMD64
-	GOOS=darwin GOARCH=amd64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) .
+	GOOS=darwin GOARCH=amd64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) ./cmd/slack-cli
 	tar -czvf $(DIST_DIR)/$(BINARY)_$(VERSION)_darwin_amd64.tar.gz -C $(DIST_DIR) $(BINARY)
 	rm $(DIST_DIR)/$(BINARY)
 
 	# Linux ARM64
-	GOOS=linux GOARCH=arm64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) .
+	GOOS=linux GOARCH=arm64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) ./cmd/slack-cli
 	tar -czvf $(DIST_DIR)/$(BINARY)_$(VERSION)_linux_arm64.tar.gz -C $(DIST_DIR) $(BINARY)
 	rm $(DIST_DIR)/$(BINARY)
 
 	# Linux AMD64
-	GOOS=linux GOARCH=amd64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) .
+	GOOS=linux GOARCH=amd64 go build $(LDFLAGS) -o $(DIST_DIR)/$(BINARY) ./cmd/slack-cli
 	tar -czvf $(DIST_DIR)/$(BINARY)_$(VERSION)_linux_amd64.tar.gz -C $(DIST_DIR) $(BINARY)
 	rm $(DIST_DIR)/$(BINARY)
