+++ cmd/slack-cli/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/piekstra/slack-cli/internal/cmd/root"
+)
+
+func main() {
+	root.Execute()
+}
