+++ internal/cmd/config/config.go
@@ -1,0 +1,19 @@
+package config
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// NewCmd creates the config command with all subcommands
+func NewCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "config",
+		Short: "Configure slack-cli",
+	}
+
+	cmd.AddCommand(newSetTokenCmd())
+	cmd.AddCommand(newDeleteTokenCmd())
+	cmd.AddCommand(newShowCmd())
+
+	return cmd
+}
