+++ internal/output/output.go
@@ -1,0 +1,5 @@
+package output
+
+// JSON is a shared flag for JSON output mode.
+// It's set by the root command and read by subcommands.
+var JSON bool
