--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/piekstra/slack-cli/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
